[#197] use separate migration folder
