+++ src/openklant/management/commands/migrate_to_v2.py
@@ -4,329 +4,24 @@
 
 import logging
 import os
-from dataclasses import asdict, dataclass, fields as dataclass_fields
-from io import BytesIO
-from typing import Any, Iterable, Optional, Tuple
+from dataclasses import fields as dataclass_fields
+from typing import Any, Tuple
 from urllib.parse import parse_qs, urlparse
 
 from django.conf import settings
 from django.core.exceptions import ImproperlyConfigured, ValidationError
 from django.core.management.base import BaseCommand, CommandParser
-from django.db.models import TextChoices
 
-import requests
-from djangorestframework_camel_case.parser import CamelCaseJSONParser, ParseError
-from djangorestframework_camel_case.render import CamelCaseJSONRenderer
 from rest_framework.fields import URLValidator
 from rest_framework.reverse import reverse
 
-from openklant.components.klantinteracties.models.constants import SoortPartij
 from openklant.components.token.models import TokenAuth
+from openklant.migration.client import Client
+from openklant.migration.v1.client import LegacyOpenKlantClient
+from openklant.migration.v1.data import Klant
+from openklant.migration.v2.client import OpenKlantClient
 
 logger = logging.getLogger(__name__)
-
-
-class Client:
-    def _request(
-        self,
-        method: str,
-        url: str,
-        data: dict | None = None,
-        params: dict | None = None,
-        headers: dict = {},
-    ) -> list | dict | None:
-        logger.debug(f"Performing {method} request on {url}")
-
-        renderer = CamelCaseJSONRenderer()
-        _data = renderer.render(data) if data else None
-
-        if method == "POST":
-            logger.debug(f"Posting: {_data}")
-            headers.update({"Content-Type": "application/json"})
-
-        try:
-            response = requests.request(
-                method, url, data=_data, params=params, headers=headers
-            )
-
-            response.raise_for_status()
-        except requests.RequestException:
-            logger.exception(f"{method} request failed for {url}")
-
-            try:
-                logger.debug(response.json())
-            except (requests.RequestException, UnboundLocalError):
-                pass
-
-            return
-
-        logger.debug(f"Received response data from {url}: {response.content}")
-
-        parser = CamelCaseJSONParser()
-
-        try:
-            response_data = parser.parse(
-                BytesIO(response.content),
-                parser_context=dict(encoding=response.encoding),
-            )
-        except ParseError:
-            logger.exception(f"Unable to parse response content: {response.content}")
-            return
-
-        return response_data
-
-    def retrieve(self, url: str, params: dict | None = None) -> list | dict | None:
-        return self._request("GET", url, params=params)
-
-    def create(self, url: str, data: dict) -> dict | None:
-        return self._request("POST", url, data=data)
-
-
-class BaseOpenKlantClient(Client):
-    token: str
-    token_prefix: str
-    base_url: str
-
-    headers: dict
-
-    def __init__(self, base_url: str, token: str) -> None:
-        self.token = token
-        self.base_url = base_url
-
-        self.headers = {"Authorization": f"{self.token_prefix} {self.token}"}
-
-    def retrieve(self, path: str, params: dict | None = None) -> list | dict | None:
-        url = self.base_url + path
-        return self._request("GET", url, headers=self.headers, params=params)
-
-    def create(self, path: str, data: dict) -> dict | None:
-        url = self.base_url + path
-        return self._request("POST", url, data=data, headers=self.headers)
-
-
-class LegacyOpenKlantClient(BaseOpenKlantClient):
-    token_prefix = "Bearer"
-
-
-class OpenKlantClient(BaseOpenKlantClient):
-    token_prefix = "Token"
-
-
-class KlantType(TextChoices):
-    natuurlijk_persoon = "natuurlijk_persoon"
-    niet_natuurlijk_persoon = "niet_natuurlijk_persoon"
-    vestiging = "vestiging"
-
-
-class Subject:
-    @property
-    def nummer(self) -> str:
-        raise NotImplementedError
-
-    def migrate(self) -> dict:
-        raise NotImplementedError
-
-
-@dataclass
-class NatuurlijkPersoon(Subject):  # SoortPartij.persoon
-    voorletters: Optional[str]
-    voornaam: Optional[str]
-    achternaam: Optional[str]
-    voorvoegsel_achternaam: Optional[str]
-
-    inp_bsn: Optional[str]
-
-    @property
-    def nummer(self) -> str:
-        return self.inp_bsn or ""
-
-    def migrate(self) -> dict:
-        return dict(
-            contactnaam=dict(
-                voorletters=self.voorletters,
-                voornaam=self.voornaam,
-                voorvoegsel_achternaam=self.voorvoegsel_achternaam,
-                achternaam=self.achternaam,
-            )
-        )
-
-
-@dataclass
-class NietNatuurlijkPersoon(Subject):  # SoortPartij.organisatie
-    inn_nnp_id: Optional[str]
-    statutaire_naam: Optional[str]
-
-    @property
-    def nummer(self) -> str:
-        return self.inn_nnp_id or ""
-
-    def migrate(self) -> dict:
-        return dict(naam=self.statutaire_naam)
-
-
-@dataclass
-class Vestiging(Subject):  # SoortPartij.organisatie
-    bedrijfsnaam: Optional[str]
-    vestigings_nummer: Optional[str]
-
-    @property
-    def nummer(self) -> str:
-        return self.vestigings_nummer or ""
-
-    def migrate(self) -> dict:
-        return dict(naam=self.bedrijfsnaam)
-
-
-class APIClass:
-    @property
-    def required_fields(self) -> Iterable:
-        raise NotImplementedError
-
-    def dict(self) -> dict:
-        return {k: v for k, v in asdict(self).items() if k in self.required_fields or v}
-
-
-@dataclass
-class DigitaalAdres(APIClass):
-    adres: str
-    soort_digitaal_adres: str = "email"
-    omschrijving: str = "Emailadres"
-    verstrekt_door_betrokkene: Optional[dict] = None
-    verstrekt_door_partij: Optional[dict] = None
-
-    @property
-    def required_fields(self):
-        return {field.name: field for field in dataclass_fields(self)}
-
-
-@dataclass
-class Partij(APIClass):
-    nummer: str
-
-    partij_identificatie: dict
-    soort_partij: str
-
-    indicatie_actief: bool = True
-    indicatie_geheimhouding: bool = False
-    rekeningnummers: Optional[list] = None
-    voorkeurs_rekeningnummer: Optional[dict] = None
-    digitale_adressen: Optional[list] = None
-    voorkeurs_digitaal_adres: Optional[dict] = None
-
-    @property
-    def required_fields(self):
-        return (
-            "digitale_adressen",
-            "voorkeurs_digitaal_adres",
-            "rekeningnummers",
-            "voorkeurs_rekeningnummer",
-            "soort_partij",
-            "indicatie_actief",
-        )
-
-
-@dataclass
-class Klant:
-    subject: Optional[str] = None
-    subject_identificatie: Optional[dict] = None
-    subject_type: Optional[str] = None
-    emailadres: Optional[str] = None
-
-    voornaam: Optional[str] = None
-    achternaam: Optional[str] = None
-    voorvoegsel_achternaam: Optional[str] = None
-
-    bedrijfsnaam: Optional[str] = None
-
-    def _get_subject(
-        self,
-    ) -> NatuurlijkPersoon | NietNatuurlijkPersoon | Vestiging | None:
-        subject_mapping = {
-            KlantType.natuurlijk_persoon: NatuurlijkPersoon,
-            KlantType.niet_natuurlijk_persoon: NietNatuurlijkPersoon,
-            KlantType.vestiging: Vestiging,
-        }
-
-        if not self.subject_type:
-            return
-
-        subject_class = subject_mapping.get(self.subject_type)
-
-        if not subject_class:
-            raise ValueError("No known subjectType found")
-
-        subject_data = self.subject_identificatie
-
-        if not isinstance(subject_data, dict) and subject_data is not None:
-            raise ValueError("Invalid subjectIdentificatie data received")
-
-        if not subject_data:
-            return
-
-        if subject_class == NatuurlijkPersoon:
-            return NatuurlijkPersoon(
-                voorletters=subject_data.get("voorletters", ""),
-                voornaam=self.voornaam,
-                achternaam=self.achternaam,
-                voorvoegsel_achternaam=self.voorvoegsel_achternaam,
-                inp_bsn=subject_data.get("inp_bsn", ""),
-            )
-        elif subject_class == NietNatuurlijkPersoon:
-            return NietNatuurlijkPersoon(
-                inn_nnp_id=subject_data.get("inn_nnp_id", ""),
-                statutaire_naam=subject_data.get("statutaire_naam", ""),
-            )
-        elif subject_class == Vestiging:
-            return Vestiging(
-                bedrijfsnaam=self.bedrijfsnaam,
-                vestigings_nummer=subject_data.get("vestigings_nummer", ""),
-            )
-
-    def _get_soort_partij(self) -> SoortPartij:
-        mapping = {
-            KlantType.natuurlijk_persoon: SoortPartij.persoon,
-            KlantType.niet_natuurlijk_persoon: SoortPartij.organisatie,
-            KlantType.vestiging: SoortPartij.organisatie,
-        }
-
-        return mapping[self.subject_type]
-
-    def to_digitaal_adres(self) -> DigitaalAdres | None:
-        if not self.emailadres:
-            return
-
-        return DigitaalAdres(adres=self.emailadres)
-
-    def to_partij(self, digitaal_adres: str | None = None) -> Partij | None:
-        try:
-            subject = self._get_subject()
-        except (TypeError, ValueError):
-            logger.exception("Unable to determine subject from subjectIdentificatie")
-            return
-
-        try:
-            soort_partij = self._get_soort_partij()
-        except KeyError:
-            logger.exception("Unable to determine soortPartij from subjectType")
-            return
-
-        data = dict(
-            indicatie_actief=True,
-            indicatie_geheimhouding=False,
-            rekeningnummers=[],
-            voorkeurs_rekeningnummer=None,
-            digitale_adressen=None,
-            voorkeurs_digitaal_adres=None,
-            soort_partij=soort_partij,
-            nummer=subject.nummer if subject else "",
-            partij_identificatie=subject.migrate() if subject else None,
-        )
-
-        if digitaal_adres:
-            data["digitale_adressen"] = [dict(uuid=digitaal_adres)]
-            data["voorkeurs_digitaal_adres"] = dict(uuid=digitaal_adres)
-
-        return Partij(**data)
 
 
 def _retrieve_klanten(url: str, access_token: str) -> Tuple[list[Klant], str | None]:
