+++ src/openklant/migration/v1/client.py
@@ -1,0 +1,5 @@
+from openklant.migration.client import BaseOpenKlantClient
+
+
+class LegacyOpenKlantClient(BaseOpenKlantClient):
+    token_prefix = "Bearer"
