+++ src/openklant/migration/v1/enum.py
@@ -1,0 +1,7 @@
+from django.db.models import TextChoices
+
+
+class KlantType(TextChoices):
+    natuurlijk_persoon = "natuurlijk_persoon"
+    niet_natuurlijk_persoon = "niet_natuurlijk_persoon"
+    vestiging = "vestiging"
