+++ src/openklant/migration/v2/client.py
@@ -1,0 +1,5 @@
+from openklant.migration.client import BaseOpenKlantClient
+
+
+class OpenKlantClient(BaseOpenKlantClient):
+    token_prefix = "Token"
